fishfry/pyproject.toml
[tool.poetry]
name = "fishfry"
version = "0.1.0"
description = ""
authors = ["Andrea Williams <andrea.k.williams7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
strawberry-graphql = {extras = ["debug-server"], version = "^0.95.1"}
quart = "^0.16.2"
hypercorn = "^0.13.2"
SQLAlchemy = "^1.4.31"
alembic = "^1.7.5"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
pytest-asyncio = "^0.17.2"
sqlalchemy2-stubs = "^0.0.2-alpha.19"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
